refactor: clean up simulation orchestration per DESIGN.md
