+++ mdfactory/orchestration/stages.py
@@ -4,8 +4,6 @@
 
 Provides :class:`StageSpec`, :data:`STAGE_REGISTRY`, :data:`STAGE_BY_NAME`,
 and the generic :func:`run_stage` that drives all four pipeline stages.
-Named convenience wrappers (:func:`run_em_stage` etc.) delegate to
-:func:`run_stage` for backward-compatible call sites.
 """
 
 from __future__ import annotations
@@ -290,162 +288,3 @@ def run_stage(
     )
 
 
-# ---- Convenience one-liner wrappers (backward-compatible named API) ----
-
-
-def run_em_stage(sim_dir: Path, grompp_app, mdrun_app, stage_config: Any = None) -> "AppFuture":
-    """Execute energy minimization stage.
-
-    Parameters
-    ----------
-    sim_dir : Path
-        Simulation directory (contains system.pdb, topology.top, em.mdp).
-    grompp_app : callable
-        Result of get_grompp_app().
-    mdrun_app : callable
-        Result of get_mdrun_app().
-    stage_config : SlurmExecutorConfig or None, optional
-        Stage-specific resource config from
-        :meth:`~mdfactory.orchestration.config.SlurmExecutorConfig.get_stage_config`.
-
-    Returns
-    -------
-    AppFuture
-        Future that resolves when mdrun completes.
-
-    """
-    return run_stage(
-        STAGE_BY_NAME["EM"], sim_dir, None, grompp_app, mdrun_app, stage_config=stage_config
-    )
-
-
-def run_nvt_stage(
-    sim_dir: Path,
-    em_future: "AppFuture | None",
-    grompp_app,
-    mdrun_app,
-    restart_from_cpt: str = "",
-    stage_config: Any = None,
-) -> "AppFuture":
-    """Execute NVT equilibration stage.
-
-    Parameters
-    ----------
-    sim_dir : Path
-        Simulation directory.
-    em_future : AppFuture or None
-        Future from EM stage (dependency). If None, assumes min.gro exists
-        (checkpoint resume scenario).
-    grompp_app : callable
-        Result of get_grompp_app().
-    mdrun_app : callable
-        Result of get_mdrun_app().
-    restart_from_cpt : str, optional
-        Path to an existing NVT checkpoint file.  When non-empty the grompp
-        step is skipped and mdrun is invoked with ``-cpi <file> -append``.
-    stage_config : SlurmExecutorConfig or None, optional
-        Per-stage resource config.
-
-    Returns
-    -------
-    AppFuture
-        Future that resolves when mdrun completes.
-
-    """
-    return run_stage(
-        STAGE_BY_NAME["NVT"],
-        sim_dir,
-        em_future,
-        grompp_app,
-        mdrun_app,
-        restart_from_cpt=restart_from_cpt,
-        stage_config=stage_config,
-    )
-
-
-def run_npt_stage(
-    sim_dir: Path,
-    nvt_future: "AppFuture | None",
-    grompp_app,
-    mdrun_app,
-    restart_from_cpt: str = "",
-    stage_config: Any = None,
-) -> "AppFuture":
-    """Execute NPT equilibration stage.
-
-    Parameters
-    ----------
-    sim_dir : Path
-        Simulation directory.
-    nvt_future : AppFuture or None
-        Future from NVT stage (dependency). If None, assumes nvt.gro and
-        nvt.cpt exist (checkpoint resume scenario).
-    grompp_app : callable
-        Result of get_grompp_app().
-    mdrun_app : callable
-        Result of get_mdrun_app().
-    restart_from_cpt : str, optional
-        Path to an existing NPT checkpoint file.  When non-empty the grompp
-        step is skipped and mdrun is invoked with ``-cpi <file> -append``.
-    stage_config : SlurmExecutorConfig or None, optional
-        Per-stage resource config.
-
-    Returns
-    -------
-    AppFuture
-        Future that resolves when mdrun completes.
-
-    """
-    return run_stage(
-        STAGE_BY_NAME["NPT"],
-        sim_dir,
-        nvt_future,
-        grompp_app,
-        mdrun_app,
-        restart_from_cpt=restart_from_cpt,
-        stage_config=stage_config,
-    )
-
-
-def run_production_stage(
-    sim_dir: Path,
-    npt_future: "AppFuture | None",
-    grompp_app,
-    mdrun_app,
-    restart_from_cpt: str = "",
-    stage_config: Any = None,
-) -> "AppFuture":
-    """Execute production MD stage.
-
-    Parameters
-    ----------
-    sim_dir : Path
-        Simulation directory.
-    npt_future : AppFuture or None
-        Future from NPT stage (dependency). If None, assumes npt.gro and
-        npt.cpt exist (checkpoint resume scenario).
-    grompp_app : callable
-        Result of get_grompp_app().
-    mdrun_app : callable
-        Result of get_mdrun_app().
-    restart_from_cpt : str, optional
-        Path to an existing Production checkpoint file.  When non-empty the
-        grompp step is skipped and mdrun is invoked with ``-cpi <file> -append``.
-    stage_config : SlurmExecutorConfig or None, optional
-        Per-stage resource config.
-
-    Returns
-    -------
-    AppFuture
-        Future that resolves when mdrun completes.
-
-    """
-    return run_stage(
-        STAGE_BY_NAME["Production"],
-        sim_dir,
-        npt_future,
-        grompp_app,
-        mdrun_app,
-        restart_from_cpt=restart_from_cpt,
-        stage_config=stage_config,
-    )
